requirements.txt
selenium==4.23.1
python-slugify==8.0.4
